% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article.id.R
\name{article.id}
\alias{article.id}
\title{A search Function}
\usage{
article.id(seedwords, originalTextFile = originalTextFile, listname = NULL)
}
\arguments{
\item{x}{Any string.}
}
\description{
This function allows you to match the seedwords to specific article.
}
\examples{
article.id()
}
\keyword{removeNumbers}
