% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del.special.characters.R
\name{del.special.characters}
\alias{del.special.characters}
\title{A preprocess Function}
\usage{
del.special.characters(x)
}
\arguments{
\item{x}{Any string.}
}
\description{
This function allows you to remove special characters and number.
}
\examples{
del.special.characters()
}
\keyword{characters}
\keyword{remove}
\keyword{special}
