% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSTClustering.R
\name{sstcjob}
\alias{sstcjob}
\title{The main Function}
\usage{
sstcjob(localfreq = 0.005, ratio = 2,
  originalTextFileName = system.file("extdata", "good_comments.txt", package =
  "SSTClustering"))
}
\arguments{
\item{localfreq}{Frequency of keywords in selected articles.}

\item{ratio}{Frequency of keywords in selected articles/All article background frequency.}

\item{originalTextFileName}{Text filename used for parsing (one article in one line),You can provide your own text, remember with absolute paths, and one line of article.}
}
\description{
This function allows you to cluster articles.
default: it use 1000 comments articles about bank services.
}
\examples{
sstcjob()
sstcjob(localfreq=0.005,ratio=2,originalTextFileName=system.file("extdata", "good_comments.txt", package = "SSTClustering"))
###Not run
sstcjob(localfreq=0.005,ratio=2,originalTextFileName="d:\\\\a.txt")
}
\keyword{sstcjob}
